Created an option to normalize the result and another option to automatically map the score value in the B-factor field of the first protein pdb. README.md has been changed to contain a quick tutorial, and two scripts have been added to the "tutorial" folder: one to generate a matrix (using gnuplot) and another to display the mapped SS3D (using vmd).

+++ Class/Atom.hpp
@@ -24,7 +24,7 @@ using namespace std;
 
 struct Atom{
 
-    int x=0, y=0, z=0;
+    long int x=0, y=0, z=0;
     char atom;
 
     Atom(){}
